06.2022/requirements.txt
loguru
hydra-core
python-logging-loki